# [SPARK-18224] [CORE] Optimise PartitionedPairBuffer implementation