--- a/core/src/main/scala/org/apache/spark/util/collection/WritablePartitionedPairCollection.scala
+++ b/core/src/main/scala/org/apache/spark/util/collection/WritablePartitionedPairCollection.scala
@@ -127,4 +127,3 @@ private[spark] trait WritablePartitionedIterator {
 
   def nextPartition(): Int
 }
-