@a-roberts 2016-11-02
Will be adding the commit from #15735 here upon addressing the feedback