@a-roberts 2016-11-02
Addressed the scalastyle comments and added the PartitionedAppendOnlyMap change here as per the above suggestions, will look at the review comments next

Two unrelated asides
1) wary I'm hogging the build machines, would be useful to not autotest everytime
2) dev/scalastyle should accept a parameter so we can quickly check just the one file, takes a long time typically