@srowen 2016-11-09
@a-roberts this does look like a worthy change, what do you think of the further simplifications here?