@a-roberts 2016-11-19
I'm resuming the work for all of these related PRs again this week after the London Spark meetup on Wednesday, if you are keen to take it on I'm more than happy to help out and will share some information here that yourself and others should find useful.

**Useful tools**
As well as simple microbenchmarking we use the [Linux perf tools](https://perf.wiki.kernel.org/index.php/Main_Page), [tprof](http://perfinsp.sourceforge.net/tprof.html) with [Visual Performance Analyzer](http://www.scalabiliti.com/library/visual_performance_analyzer) and also IBM's [Healthcenter](https://www.ibm.com/developerworks/java/jdk/tools/healthcenter/) for Java for method profiling (this is bundled with the JDK and you provide -Xhealthcenter as a driver/executor option then open the files in said tool).

**Benchmarks**
We'd want to run our improvement ideas with and without the changes using[ HiBench 6](https://github.com/intel-hadoop/HiBench) (large profile) and [SparkSqlPerf](https://github.com/databricks/spark-sql-perf) against all 100 TPCDS queries.