@a-roberts 2016-11-25
Back to working on the performance related JIRAs now, so based on the above helpful comments here's what I'll do

Remove the .get.compare from the loop as suggested above - we'll do a .get  upfront to get our comparator to use, eliminating the .get later

Move the duplicated code into the WritablePartitionedPairCollection object so the two methods optimised here will call the above new method (let's say it's called getComparator) before returning accordingly (both methods are the same apart from the final few lines).

PartitionedAppendOnlyMap returns 
```
destructiveSortedIterator(comparator)
```

and PartitionedPairBuffer returns:

```
new Sorter(new KVArraySortDataFormat[(Int, K), AnyRef]).sort(data, 0, curSize, comparator)
iterator
```

I'll then build/test/profile this again