@srowen 2016-11-29
If that's true then again doesn't my suggestion to inline `partitionComparator` fix it?